validacoes simples nos request

+++ microservice_1/app/bussines/Users.py
@@ -18,6 +18,13 @@ def get_all_users():
 @bp_users.route('/user/<int:id>', methods=['GET'])
 def get_user(id):
     user = User.query.get(id)
+
+    if user == None:
+        return jsonify({
+            'success': False,
+            'message': 'Usuário inexistente'
+        }), 401
+
     result = user_schema.dump(user)
 
     return jsonify({'users': result})
@@ -26,10 +33,17 @@ def get_user(id):
 def create_new_user():
     data_user_json = user_schema.load(request.json)
 
+    # Checa se todos os dados foram informados e se não tem não existem dados nulos
+    if len(data_user_json) < 3 or "" in data_user_json.values():
+        return jsonify({
+            'success': False,
+            'message': 'Você deve informar todos os dados!'
+        }), 401
+
     email = data_user_json['email']
     username = data_user_json['username']
     password = data_user_json['password']
-    
+
     new_user = User(email, username, password)
 
     current_app.db.session.add(new_user)
@@ -41,7 +55,19 @@ def create_new_user():
 def update_user(id):
     user = User.query.get(id)
 
+    if user == None:
+        return jsonify({
+            'success': False,
+            'message': 'Usuário inexistente'
+        }), 401
+
     data_user_json = user_schema.load(request.json)
+
+    if len(data_user_json) < 3 or '' in data_user_json.values():
+        return jsonify({
+            'success': False,
+            'message': 'Informe todos os dados'
+        }), 401
 
     user.email = data_user_json['email']
     user.username = data_user_json['username']
@@ -54,6 +80,12 @@ def update_user(id):
 @bp_users.route('/user/<int:id>', methods=['DELETE'])
 def delete_user(id):
     user = User.query.get(id)
+    
+    if user == None:
+        return jsonify({
+            'success': False,
+            'message': 'Usuário inexistente'
+        }), 401
 
     current_app.db.session.delete(user)
     current_app.db.session.commit()
